+++ src/api/v1/routes/board/boards.ts
@@ -12,7 +12,7 @@ import { eq, and, inArray, exists } from "drizzle-orm";
 import {
   type WithSessionVariables,
   USER,
-} from "../auth/utils/contextSessionVaraibles";
+} from "../auth/utils/contextSessionVariables";
 import { z } from "zod";
 import { StatusCodes } from "http-status-codes";
 import { getApiErrorShape } from "../../../../db/v1/utils/apiGeneralTypes";