+++ src/api/v1/routes/organization/member/organizationMembers.ts
@@ -1,0 +1,5 @@
+import { Hono } from "hono";
+
+const organizationMembers = new Hono().basePath("/members");
+
+export { organizationMembers };